requirements.txt
numpy
scipy
pandas
librosa
scikit-learn
matplotlib
soundfile
requests
pyyaml
librosa==0.11.0
numpy==1.26.4
soundfile==0.13.1
